Fixed conversation opening

+++ database/conversations/member.go
@@ -8,8 +8,8 @@ type Member struct {
 	Conversation uint `json:"conversation" gorm:"not null"`
 
 	// 1 - member, 2 - admin, 3 - owner
-	Role    uint `json:"role" gorm:"not null"`
-	Account uint `json:"account" gorm:"not null"`
+	Role    uint  `json:"role" gorm:"not null"`
+	Account int64 `json:"account" gorm:"not null"`
 
 	// Relationships
 	Status fetching.Status `json:"-" gorm:"foreignKey:Account"`
